Add new makefile
Now it easier to manage the makefile, to add a new file to the compilation
simply add the "file.o" to the "_OBJ" variable of the makefile.

makefile was moved to the root of the directory for easier access.

For compiling the debug version just use: make debug
And for changing the compiler: make CC=clang++