Port to windows
Now the linux exclusive parts of the code are guarded by __linux and an
alternative code for windows is implement on the else.

It is been compiling for windows using cygwin.

The alternative is not equivalent:
linux: check for all files on the watched directory using inotify
windows: poll the folder checking for the new file, and if exists send
the previous file (to make sure the file is closed).